Add files via upload
The screen where user selects a song will implement this code, it takes a given path to a folder of mp3 files and gets the name of each song. The name is then displayed on the buttons. Subject to change if we use an API for music data instead of a folder with a fixed song collection.